executor: turn on O_CLOEXEC for all deserialized fds
We have this rule in systemd that all our fds should always have
O_CLOEXEC set, except were they really should not. Hence, when taking
possession of the deseiralized fds turn on O_CLOEXEC. We'll later turn
it off again for the specific ones we want to pass along.

Alternative for: #30271 #30266